+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Version 0.1.8 2022-01-02
+
+ * Allow Google Storage blob (with public URLs)
+ * Remove Travis CI build
+
+## Version 0.1.7 2021-03-14
+
+ * Allow file-like objects
 ## Version 0.1.2
 
  * Implemented pull request #3, thanks @sih4sing5hong5